Add directory structure and starter files
These are all linked together, so I'm adding them all at once.

Top level index is just an entry point.

In src is the source code. discord and mongodb files are
both there to connect to their respective services. Index
actually calls start-fxns from them to cause bot to log in.

In handlers, a manager determines who gets what (to simplify
discord file in above directory). Handlers live in subdirectories.
These are more proof-of-concept than finalized names/handlers.
Also, setup-handler is missing.

In util are constants and a console wrapper so we can turn off
debug if we want, etc.