Created Confirms for Init Tourney
Created the backend of message confirmations, in
a new mongo file that has its own schema and whatnot.
Simply stores a message by ID so we can confirm
that message reactions are on a relevant emoji.

Created confirmation message send and receive for
asking if someone meant to init a tournament.

YAY IT WORKS