Added payload to messages, moved receive to util
By adding a payload, I can return team names, data, and
more. This is to simplify what comes out of the confirm.

Also I learned that promises can only return one item, so
the return is now an object with a "status" and a "payload".

The yes-no logic for receiving was moved to util to minimize
code duplication.