Removed inherent bug from DM Channels
Problem: if you dm '@bot', it tries to access the guild
which doesn't exist in a dm channel and then it dies.

(dm = direct message)

Solution: check if a message has a guild object; if not,
reply that the bot isn't equipped to handle such a message
and quit before any calls asking for the guild.