Remove messages from the database after confirm
You could react to and trip bot responses by reacting
multiple times to one message, even with opposite reactions.

This PR fixes that by clearing out messages as they are responded
to. You can only respond once to Y/N confirmation messages.

Consideration: dispute messages are still the same
as they are, and they will never be removed from the database
as is. This might be a serious flaw in large-scale deployment
because the db gets clogged with useless messages.