Add participant progress (#126)
* pulling

* starting to work with errors

* added staged tourney schema

* workign through chat handler and guild staging

* added set team option, set team name and getters. added chat state handler backend for dialogue and movign from init to setup

* added tournament player cap to db as well as check in setup tourney handler

* added leftover tourneytyeptostring

* fixed err handler return

* added team allowed check to setup tournye handler

* Update manager.js

fixed ===

* Update handler.js

changed fulfill to not bug out chat state

* Update handler.js

* Update manager.js

reverting a change, sorry on phone >.>

* Update chat_state_handle.js

changed to align with latest parser

* Groundwork for disputes

* Tournament resoltuion is working!

Finalize challonge, retreive winner, and advance tournament state in process_winner
TODO: Post name to Discord
Close tourney isn't working, so I'm fixing that now

* Close tourney now works too!

We may want to refine it at some point, but when the tournament finishes:
The winner is retreived and logged, the tournament is stashed, and the tournament
state is advanced so DB cleanup starts. When you ask the bot to close the tourney,
it clears all the channels.

* WIP: Handler still not working, but pulling in updated parser code
to try and fix it.

* Disputes are finally working!

Had an issue where due to a merge conflict my branch fractured,
so my last committ didn't go to a branch. I went back through and
copied the changes from a back-up file, but when I re-based from
master it pulled in Alexei's most recent changes and I can no longer
run the bot. Wanted to get this commit in now before my branch becomes
corrupted again!

TODO: Send messages to the right channels upon victory, tournament ending,
dispute resolving, etc.
TODO: Update vote threshold, currently using test comparison

* fixed chat state integration -- WIP!

* fixed remaining chat state errors

* EchoBots only reply if @bot is first word

* TourneyAnounce updates with winner and stashed url

* Roles are now hoisted in the sidebar correctly

Now you visibly see yourself added to teams on the sidebar.

Also created a test function in the tester bot.

* init parser and errors first draft

* Added foundation for sending messages to channels

* Added match, dispute, and victory messages

* Added the final set of run_tourney messages!

* init parsing workaround error

* fixed chat state handler to be more modular, added constant STAGED_PROPS (needed since staged_t has a lot of random props)

* about to incorporate workaround changes

* init parsing workaround error (#94)

merging this so I can debug it . using sqash + merge so its quickly reversible

* starting to work with errors

* init parser and errors first draft

* init parsing workaround error

* about to incorporate workaround changes

* reverting master

* Revert "init parsing workaround error (#94)" (#96)

This reverts commit 3755abd53bb79ffde1ce6b82c419581579212580.

* debugged parser issues blocking init

* fixed init nlp

* added confirm flag to init

* Say hello when entering a server

* added resolve

* tourney init asks for confirm before pushing forward

* Bot now initiates the conversation when you make a tournament

* Added notifying functions

In webservices/discord_notifications, there are fxns
wrapping notifications for users/roles.

* Removed inherent bug from DM Channels

Problem: if you dm '@bot', it tries to access the guild
which doesn't exist in a dm channel and then it dies.

(dm = direct message)

Solution: check if a message has a guild object; if not,
reply that the bot isn't equipped to handle such a message
and quit before any calls asking for the guild.

* Oops semicolon

* Added comments to notifications explaining returns

* Added description of how we test to contribute.md

This way, devs who read this can test like we do. Yup.

It just describes how to set up and use the echobots.

* Exit process if mongodb isn't running

Also added newline at end of file. :-P

* Give bot permission to talk in init-channel

* Fixed permissions bugs

* Added note about reactions to bot entry message

* Added tournament-making test file

By running node src/test/make_t.js, you create a
Challonge tournament using the API, so you can
tell if the clear script actually deletes a tournament
on Challonge.

* Made wipeDB implement promises

It looked like callback hell so I rearranged it.

It's functionally identical, but promise-chaining is the
whole point of promises.

* Added wipe-msg-db to cleardb command

The message database was separate from everything, so
now I've added it to the cleardb command to make sure
the back end is totally cleaned up by cleardb. :-)

* error handling progress

* Remove messages from the database after confirm

You could react to and trip bot responses by reacting
multiple times to one message, even with opposite reactions.

This PR fixes that by clearing out messages as they are responded
to. You can only respond once to Y/N confirmation messages.

Consideration: dispute messages are still the same
as they are, and they will never be removed from the database
as is. This might be a serious flaw in large-scale deployment
because the db gets clogged with useless messages.

* Added helper and inserted into match headers

Now you can enter a role id with a guild object to get a

@role (@player @player)

output.

* Testing functions update

* added gcloud app.yaml file

* added new mongodb adress

* added gcloud cmds to circle yml

* added stop prev ver to gcloud args

* Create README.md

got rid of extra circleci badge

* added -q app.yaml to gcloud cmd

* removed python env line

* added sudo to circle yaml

* switched all cmds to cmds

* removed env from yaml

* im gonna kill the ,maker of yaml files

* guild schema correction

* fixed lint errors in parse

* addded sudo to gcloud cmds

* added bin sudo

* got rid of app.yaml in gcloud deploy

* testing hardcoding db for cicleci deploy

* Added t-player name to constants

* Added general role for t-players

* removed req creds

* changed version

* removed deployment from circle

* test circle w/o quite

* Added notifyAllPlayers function

* testing beta on circle

* circle

* working on error handler- join tourney edition

* added unlimited text

* removed deploy from circle

* Stopped missing createDM from killing bot

* Attempted fix of dms

* error handler nearly complete. Just need to work on not allowing participants to join additional teams

* fixed cap=0 bug

* Fixed dispute threshold to work off of num_teams_remaining

Removes losing team after each match is over to reduce num_teams

* working on join team bug

* Updated banner messages for every channel to add helpful instructions.

* Made match confirmations reflect the role_id (aka team name)

* fixed findParticipantTeamID bug

* Switched back to local Mongo

* fixed joining bug